moved energyDataApi to energyDataService - it is more Angulary!
