+++ src/angular2/AngularSolarEnergy/src/app/app.module.ts
@@ -1,4 +1,3 @@
-import EnergyDataApi from './energy/energyDataApi';
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
@@ -6,6 +5,13 @@ import { HttpModule } from '@angular/http';
 
 import { AppComponent } from './app.component';
 import { EnergyComponent } from './energy/energy.component';
+import EnergyDataService from './energy/energy-data.service';
+
+// Imports for loading & configuring the in-memory web api
+// https://angular.io/docs/ts/latest/tutorial/toh-pt6.html
+// import { InMemoryWebApiModule } from 'angular-in-memory-web-api';
+// import { InMemoryDataService }  from './in-memory-data.service';
+
 
 @NgModule({
   declarations: [
@@ -17,7 +23,7 @@ import { EnergyComponent } from './energy/energy.component';
     FormsModule,
     HttpModule
   ],
-  providers: [EnergyDataApi],
+  providers: [EnergyDataService],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
