+++ src/angular2/AngularSolarEnergy/src/app/energy/energyRowViewModel.ts
